techrag-ingester/requirements.txt
graphiti-core[neo4j,openai]
python-dotenv
openai